2010-07-17  Daniel Franke  <dfranke@users.sourceforge.net>
	* saxsimage_format.h (saxs_image_format): Changed member
	function-pointer 'value' to return a signed value, same for
	'value_min' and 'value_max'.
	* saxsimage.c (image_value_max): Removed unnecessary check.
	* tiff.c (saxs_image_tiff_value): Changed return value type
	to signed.
	* cbf.c (saxs_image_cbf_open): Initialize private data members.
	(saxs_image_cbf_read_high_level): Read signed values.
	(saxs_image_cbf_value): Changed return value type to signed.

+++ libsaxsimage/saxsimage_format.h
@@ -38,11 +38,11 @@ struct saxs_image_format {
   int (*write)(void *private_data, const char *filename);
   int (*close)(void *private_data);
 
-  size_t (*value)(void *private_data, int x, int y);
+  long (*value)(void *private_data, int x, int y);
   size_t (*width)(void *private_data);
   size_t (*height)(void *private_data);
-  size_t (*value_min)(void *private_data);
-  size_t (*value_max)(void *private_data);
+  long (*value_min)(void *private_data);
+  long (*value_max)(void *private_data);
 };
 typedef struct saxs_image_format saxs_image_format;
